+++ tests/conftest.py
@@ -15,15 +15,13 @@ You should have received a copy of the GNU General Public License
 along with AIMMD. If not, see <https://www.gnu.org/licenses/>.
 """
 import pytest
-# TODO/BUG: weird stuff: if not importing mdtraj before RCModel
-# the tests segfault....!?
-import mdtraj
 import numpy as np
 import openpathsampling as paths
 import openpathsampling.engines.toy as toys
 from openpathsampling.engines import Trajectory as OPSTrajectory
 from functools import reduce
-from aimmd.base.rcmodel import RCModel
+
+from helpers import OneOutRCModel, TwoOutRCModel
 
 
 def pytest_addoption(parser):
@@ -61,6 +59,36 @@ def pytest_collection_modifyitems(config, items):
             item.add_marker(skip_slow)
         if not old and "old" in item.keywords:
             item.add_marker(skip_old)
+
+
+@pytest.fixture
+def oneout_rcmodel():
+    return OneOutRCModel(lambda x: -x)
+
+
+@pytest.fixture
+def oneout_rcmodel_notrans():
+    return OneOutRCModel(None)
+
+
+@pytest.fixture
+def oneout_rcmodel_opstrans():
+    def transform(x):
+        if isinstance(x, OPSTrajectory):
+            return np.array([[0.]])
+        else:
+            return np.array([[-200.]])
+    return OneOutRCModel(transform)
+
+
+@pytest.fixture
+def twoout_rcmodel():
+    return TwoOutRCModel(lambda x: -x)
+
+
+@pytest.fixture
+def twoout_rcmodel_notrans():
+    return TwoOutRCModel(None)
 
 
 @pytest.fixture
@@ -132,85 +160,3 @@ def ops_toy_sim_setup():
             'modifier': modifier,
             'cv_ndim': n_harmonics + 2,
             }
-
-
-@pytest.fixture
-def oneout_rcmodel():
-    return OneOutRCModel(lambda x: -x)
-
-
-@pytest.fixture
-def oneout_rcmodel_notrans():
-    return OneOutRCModel(None)
-
-
-@pytest.fixture
-def oneout_rcmodel_opstrans():
-    def transform(x):
-        if isinstance(x, OPSTrajectory):
-            return np.array([[0.]])
-        else:
-            return np.array([[-200.]])
-    return OneOutRCModel(transform)
-
-
-@pytest.fixture
-def twoout_rcmodel():
-    return TwoOutRCModel(lambda x: -x)
-
-
-@pytest.fixture
-def twoout_rcmodel_notrans():
-    return TwoOutRCModel(None)
-
-
-# Test helper classes
-class OneOutRCModel(RCModel):
-    def __init__(self, transform=None):
-        # work on numpy arrays directly, but test the use of transform
-        # using transform should exchange p_A with p_B
-        super().__init__(states=["A", "B"], descriptor_transform=transform)
-        self.n_train = 0
-
-    @property
-    def n_out(self):
-        return 1
-
-    def train_hook(self, trainset):
-        # just to have something to check if
-        # and how often we would have trained
-        self.n_train += 1
-
-    def _log_prob(self, descriptors, batch_size):
-        # so we have a return value that has the correct shape
-        return np.sum(descriptors, axis=1, keepdims=True)
-
-    def test_loss(self, trainset, batch_size):
-        # just to be able to instantiate
-        raise NotImplementedError
-
-
-class TwoOutRCModel(RCModel):
-    def __init__(self, transform):
-        super().__init__(states=["A", "B"], descriptor_transform=transform)
-        self.n_train = 0
-
-    @property
-    def n_out(self):
-        return 2
-
-    def train_hook(self, trainset):
-        # just to have something to check if
-        # and how often we would have trained
-        self.n_train += 1
-
-    def _log_prob(self, descriptors, batch_size):
-        # so we have a return value that has the correct shape
-        q_A = np.sum(descriptors, axis=1, keepdims=True)
-        # take q_B = -q_A such that p_A + p_B = 1
-        q_B = -np.sum(descriptors, axis=1, keepdims=True)
-        return np.concatenate((q_A, q_B), axis=1)
-
-    def test_loss(self, trainset, batch_size):
-        # just to be able to instantiate
-        raise NotImplementedError
