+++ tests/helpers.py
@@ -1,0 +1,73 @@
+# This file is part of aimmd
+#
+# aimmd is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# aimmd is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with aimmd. If not, see <https://www.gnu.org/licenses/>.
+"""
+This file contains helper classes for tests.
+
+Usually these will be provided via fixtures created in conftest.py
+"""
+import numpy as np
+from aimmd.base.rcmodel import RCModel
+
+
+# Test helper classes
+class OneOutRCModel(RCModel):
+    def __init__(self, transform=None):
+        # work on numpy arrays directly, but test the use of transform
+        # using transform should exchange p_A with p_B
+        super().__init__(states=["A", "B"], descriptor_transform=transform)
+        self.n_train = 0
+
+    @property
+    def n_out(self):
+        return 1
+
+    def train_hook(self, trainset):
+        # just to have something to check if
+        # and how often we would have trained
+        self.n_train += 1
+
+    def _log_prob(self, descriptors, batch_size):
+        # so we have a return value that has the correct shape
+        return np.sum(descriptors, axis=1, keepdims=True)
+
+    def test_loss(self, trainset, batch_size):
+        # just to be able to instantiate
+        raise NotImplementedError
+
+
+class TwoOutRCModel(RCModel):
+    def __init__(self, transform):
+        super().__init__(states=["A", "B"], descriptor_transform=transform)
+        self.n_train = 0
+
+    @property
+    def n_out(self):
+        return 2
+
+    def train_hook(self, trainset):
+        # just to have something to check if
+        # and how often we would have trained
+        self.n_train += 1
+
+    def _log_prob(self, descriptors, batch_size):
+        # so we have a return value that has the correct shape
+        q_A = np.sum(descriptors, axis=1, keepdims=True)
+        # take q_B = -q_A such that p_A + p_B = 1
+        q_B = -np.sum(descriptors, axis=1, keepdims=True)
+        return np.concatenate((q_A, q_B), axis=1)
+
+    def test_loss(self, trainset, batch_size):
+        # just to be able to instantiate
+        raise NotImplementedError
